+++ src/calendar/components/WeekBar.js
@@ -1,14 +1,28 @@
-import React from 'react';
+import React, {PureComponent} from 'react';
 import {View, Text} from 'react-native';
 import PropTypes from 'prop-types';
-import {weekBarStyles} from "../style";
+import {weekBarStyles} from '../style';
 
-const WeekBar = ({weeks, style, textStyle}) => (
-    <View style={[weekBarStyles.view, style]}>
-        {weeks.map((week, index) => <Text key={index} style={[weekBarStyles.text, textStyle]}>{week}</Text>)}
-    </View>
-);
+// Uses PureComponent to improve performance.
+class WeekBar extends PureComponent {
 
+    render() {
+
+        const {
+            weeks,
+            style,
+            textStyle,
+        } = this.props;
+
+        return (
+            <View style={[weekBarStyles.view, style]}>
+                {weeks.map((week, index) => <Text key={index} style={[weekBarStyles.text, textStyle]}>{week}</Text>)}
+            </View>
+        );
+    }
+}
+
+// Props defined here.
 WeekBar.propTypes = {
     weeks: PropTypes.array.isRequired,
     style: PropTypes.any,